+++ src/Communication/Race/dataInterfaces.ts
@@ -1,5 +1,5 @@
 export interface ItemUsedEvent {
 	itemType: string;
 	targetPlayerId: string;
-	fromPayerId: string;
+	fromPayerId?: string;
 }
