--- src/Communication/SocketEvent.ts
@@ -1,4 +1,0 @@
-export default interface SocketEvent {
-	readonly Name: string;
-	data: any;
-}
