--- src/GameCore/GAMECORE_CST.ts
@@ -1,6 +1,0 @@
-export namespace GAMECORE_CST {
-	export const enum Norm {
-		Euclidian = "Euclidian",
-		Taxicab = "Taxicab",
-	}
-}
